fix: update export type
Export factory functions for the transport and peer discovery separately
