+++ packages/webrtc-star-transport/test/compliance.spec.ts
@@ -9,12 +9,13 @@ import testsDiscovery from '@libp2p/interface-peer-discovery-compliance-tests'
 import { webRTCStar } from '../src/index.js'
 import pWaitFor from 'p-wait-for'
 import { peerIdFromString } from '@libp2p/peer-id'
+import type { WebRTCStar } from '../src/transport.js'
 
 describe('interface-transport compliance', function () {
   testsTransport({
     async setup () {
       const peerId = peerIdFromString('QmcgpsyWgH8Y8ajJz1Cu72KnS5uo2Aa2LpzU7kinSooo2a')
-      const ws = webRTCStar({ wrtc })({ peerId })
+      const ws = webRTCStar({ wrtc }).transport({ peerId })
 
       const base = (id: string) => {
         return `/ip4/127.0.0.1/tcp/15555/ws/p2p-webrtc-star/p2p/${id}`
@@ -47,7 +48,7 @@ describe('interface-discovery compliance', () => {
   testsDiscovery({
     async setup () {
       const peerId = peerIdFromString('QmcgpsyWgH8Y8ajJz1Cu72KnS5uo2Aa2LpzU7kinSooo2d')
-      const ws = webRTCStar({ wrtc })({ peerId })
+      const ws = webRTCStar({ wrtc }).transport({ peerId }) as WebRTCStar
       const maStr = '/ip4/127.0.0.1/tcp/15555/ws/p2p-webrtc-star/p2p/QmcgpsyWgH8Y8ajJz1Cu72KnS5uo2Aa2LpzU7kinSooo2d'
 
       const discovery = ws.discovery()
