+++ packages/webrtc-star-transport/test/node.ts
@@ -15,6 +15,7 @@ import trackTests from './transport/track.js'
 import reconnectTests from './transport/reconnect.node.js'
 import type { PeerTransport } from './index.js'
 import { mockRegistrar, mockUpgrader } from '@libp2p/interface-mocks'
+import type { WebRTCStar, WebRTCStarDiscovery } from '../src/transport.js'
 
 // TODO: Temporary fix per wrtc issue
 // https://github.com/node-webrtc/node-webrtc/issues/636#issuecomment-774171409
@@ -23,19 +24,22 @@ process.on('beforeExit', (code) => process.exit(code))
 describe('transport: with wrtc', () => {
   const create = async (): Promise<PeerTransport> => {
     const peerId = await createEd25519PeerId()
-    const ws = webRTCStar({
-      wrtc
-    })({ peerId })
+    const wrtcStar = webRTCStar({ wrtc })
+    const transport = wrtcStar.transport({ peerId }) as WebRTCStar
+    const discovery = wrtcStar.discovery() as WebRTCStarDiscovery
 
     const registrar = mockRegistrar()
     const upgrader = mockUpgrader({ registrar })
 
-    return {
+    const peerTransport: PeerTransport = {
       peerId,
-      transport: ws,
+      transport,
+      discovery,
       registrar,
       upgrader
     }
+
+    return peerTransport
   }
 
   dialTests(create)
@@ -51,19 +55,22 @@ describe('transport: with wrtc', () => {
 describe.skip('transport: with electron-webrtc', () => {
   const create = async () => {
     const peerId = await createEd25519PeerId()
-    const ws = webRTCStar({
-      wrtc: electronWebRTC()
-    })({ peerId })
+    const wrtcStar = webRTCStar({ wrtc: electronWebRTC() })
+    const transport = wrtcStar.transport({ peerId }) as WebRTCStar
+    const discovery = wrtcStar.discovery() as WebRTCStarDiscovery
 
     const registrar = mockRegistrar()
     const upgrader = mockUpgrader({ registrar })
 
-    return {
+    const peerTransport: PeerTransport = {
       peerId,
-      transport: ws,
+      transport,
+      discovery,
       registrar,
       upgrader
     }
+
+    return peerTransport
   }
 
   dialTests(create)
