+++ packages/webrtc-star-transport/test/transport/dial.ts
@@ -8,7 +8,7 @@ import { fromString as uint8ArrayFromString } from 'uint8arrays/from-string'
 import sinon from 'sinon'
 import { WebRTCReceiver } from '@libp2p/webrtc-peer'
 import { cleanUrlSIO } from '../../src/utils.js'
-import type { WebRTCStar } from '../../src/index.js'
+import type { WebRTCStar } from '../../src/transport.js'
 import type { Listener, Upgrader } from '@libp2p/interface-transport'
 import pWaitFor from 'p-wait-for'
 import type { HandshakeSignal } from '@libp2p/webrtc-star-protocol'
