+++ packages/webrtc-star-transport/test/transport/discovery.ts
@@ -5,7 +5,7 @@ import { multiaddr } from '@multiformats/multiaddr'
 import type { Multiaddr } from '@multiformats/multiaddr'
 import { pEvent } from 'p-event'
 import { cleanUrlSIO } from '../../src/utils.js'
-import type { WebRTCStar } from '../../src/index.js'
+import type { WebRTCStar } from '../../src/transport.js'
 import type { Listener } from '@libp2p/interface-transport'
 import { mockUpgrader } from '@libp2p/interface-mocks'
 import type { PeerTransport } from '../index.js'
