+++ packages/webrtc-star-transport/test/transport/instance.spec.ts
@@ -5,8 +5,13 @@ import { expect } from 'aegir/chai'
 import { webRTCStar } from '../../src/index.js'
 
 describe('instantiate the transport', () => {
-  it('create', async () => {
-    const wstar = webRTCStar()({ peerId: await createEd25519PeerId() })
+  it('create transport', async () => {
+    const wstar = webRTCStar().transport({ peerId: await createEd25519PeerId() })
+    expect(wstar).to.exist()
+  })
+
+  it('create discovery', async () => {
+    const wstar = webRTCStar().discovery({ peerId: await createEd25519PeerId() })
     expect(wstar).to.exist()
   })
 })
