+++ packages/webrtc-star-transport/test/transport/track.ts
@@ -7,7 +7,7 @@ import type { Multiaddr } from '@multiformats/multiaddr'
 import { pipe } from 'it-pipe'
 import pWaitFor from 'p-wait-for'
 import { cleanUrlSIO } from '../../src/utils.js'
-import type { WebRTCStar } from '../../src/index.js'
+import type { WebRTCStar } from '../../src/transport.js'
 import type { Listener, Upgrader } from '@libp2p/interface-transport'
 import { mockRegistrar, mockUpgrader } from '@libp2p/interface-mocks'
 import type { PeerTransport } from '../index.js'
